+++ native/src/test/java/io/ballerina/stdlib/sql/nativeimpl/OutParameterProcessorTest.java
@@ -16,16 +16,15 @@
  *  under the License.
  */
 
-package io.ballerina.stdlib.sql.tests.nativeimpl;
+package io.ballerina.stdlib.sql.nativeimpl;
 
 import io.ballerina.runtime.api.PredefinedTypes;
 import io.ballerina.runtime.api.values.BMap;
 import io.ballerina.runtime.api.values.BObject;
 import io.ballerina.runtime.api.values.BString;
 import io.ballerina.stdlib.sql.Constants;
-import io.ballerina.stdlib.sql.nativeimpl.OutParameterProcessor;
+import io.ballerina.stdlib.sql.TestUtils;
 import io.ballerina.stdlib.sql.parameterprocessor.DefaultResultParameterProcessor;
-import io.ballerina.stdlib.sql.tests.TestUtils;
 import org.testng.annotations.Test;
 
 import static io.ballerina.runtime.api.utils.StringUtils.fromString;
