+++ native/src/test/java/io/ballerina/stdlib/sql/parameterprocessor/DefaultResultParameterProcessorTest.java
@@ -16,7 +16,7 @@
  *  under the License.
  */
 
-package io.ballerina.stdlib.sql.tests.parameterprocessor;
+package io.ballerina.stdlib.sql.parameterprocessor;
 
 import io.ballerina.runtime.api.PredefinedTypes;
 import io.ballerina.runtime.api.utils.TypeUtils;
@@ -25,10 +25,9 @@ import io.ballerina.runtime.api.values.BMap;
 import io.ballerina.runtime.api.values.BObject;
 import io.ballerina.runtime.api.values.BString;
 import io.ballerina.runtime.internal.values.MapValueImpl;
+import io.ballerina.stdlib.sql.TestUtils;
 import io.ballerina.stdlib.sql.exception.ApplicationError;
 import io.ballerina.stdlib.sql.exception.DataError;
-import io.ballerina.stdlib.sql.parameterprocessor.DefaultResultParameterProcessor;
-import io.ballerina.stdlib.sql.tests.TestUtils;
 import io.ballerina.stdlib.sql.utils.PrimitiveTypeColumnDefinition;
 import org.testng.annotations.Test;
