+++ native/src/test/java/io/ballerina/stdlib/sql/utils/UtilsTest.java
@@ -16,13 +16,11 @@
  *  under the License.
  */
 
-package io.ballerina.stdlib.sql.tests.utils;
+package io.ballerina.stdlib.sql.utils;
 
 import io.ballerina.runtime.api.types.StructureType;
 import io.ballerina.runtime.api.utils.TypeUtils;
-import io.ballerina.stdlib.sql.tests.TestUtils;
-import io.ballerina.stdlib.sql.utils.PrimitiveTypeColumnDefinition;
-import io.ballerina.stdlib.sql.utils.Utils;
+import io.ballerina.stdlib.sql.TestUtils;
 import org.testng.annotations.Test;
 
 import java.util.ArrayList;
